move and change the path of the components to their respective folders as part of enhancement #26
